+++ lib/mix/lib/mix/tasks/compile.elixir.ex
@@ -19,8 +19,12 @@ defmodule Mix.Tasks.Compile.Elixir do
   force compilation regardless of mod times by passing
   the `--force` option.
 
-  A list of files can be given after the task
-  name in order to select the files to compile.
+  Note it is important to recompile all files because
+  often there are compilation time dependencies between
+  the files (macros and etc). However, in some cases it
+  is useful to compile just the changed files for quick
+  development cycles, for such, a developer can pass
+  the `--quick` otion.
 
   ## Configuration
 
@@ -31,38 +35,47 @@ defmodule Mix.Tasks.Compile.Elixir do
 
   ## Command line options
 
-  * `--force` - forces compilation regardless of mod times;
+  * `--force` - forces compilation regardless of module times;
+  * `--quick`, `-q` - only compile files that changed;
 
   """
   def run(args) do
-    { opts, files } = OptionParser.parse(args, flags: [:force])
+    { opts, _ } = OptionParser.parse(args,
+                    flags: [:force, :quick], aliases: [q: :quick])
 
     project       = Mix.project
     compile_path  = project[:compile_path]
-    compile_first = project[:compile_first]
     compile_exts  = project[:compile_exts]
     watch_exts    = project[:watch_exts]
     source_paths  = project[:source_paths]
 
-    to_compile = Mix.Utils.extract_files(source_paths, files, compile_exts)
-    to_watch   = Mix.Utils.extract_files(source_paths, files, watch_exts)
+    to_compile = Mix.Utils.extract_files(source_paths, compile_exts)
+    to_watch   = Mix.Utils.extract_files(source_paths, watch_exts)
+    stale      = Mix.Utils.extract_stale(to_watch, [compile_path])
 
-    if opts[:force] or Mix.Utils.stale?(to_watch, [compile_path]) do
+    if opts[:force] or stale != [] do
       File.mkdir_p! compile_path
-      Code.delete_path compile_path
-
-      if elixir_opts = project[:elixirc_options] do
-        Code.compiler_options(elixir_opts)
-      end
-
-      ordered = List.uniq compile_first ++ to_compile
-      compile_files ordered, compile_path
-
-      Code.prepend_path compile_path
+      compile_files opts[:quick], project, compile_path, to_compile, stale
       :ok
     else
       :noop
     end
+  end
+
+  defp compile_files(true, project, compile_path, _to_compile, stale) do
+    opts = project[:elixirc_options] || []
+    opts = Keyword.put(opts, :ignore_module_conflict, true)
+    Code.compiler_options(opts)
+    compile_files stale, compile_path
+  end
+
+  defp compile_files(false, project, compile_path, to_compile, _stale) do
+    Code.delete_path compile_path
+    if elixir_opts = project[:elixirc_options] do
+      Code.compiler_options(elixir_opts)
+    end
+    compile_files to_compile, compile_path
+    Code.prepend_path compile_path
   end
 
   defp compile_files(files, to) do